requirements.txt
asgiref==3.2.7
certifi==2020.4.5.1
chardet==3.0.4
defusedxml==0.6.0
Django==3.0.5
django-allauth==0.41.0
django-cors-middleware==1.5.0
django-oauth-toolkit==1.3.2
django-redis==4.11.0
django-rest-auth==0.9.5
djangorestframework==3.11.0
idna==2.9
oauthlib==3.1.0
psycopg2==2.8.5
pycryptodome==3.9.7
python3-openid==3.1.0
pytz==2020.1
redis==3.5.0
requests==2.23.0
requests-oauthlib==1.3.0
six==1.14.0
sqlparse==0.3.1
urllib3==1.25.9
uWSGI==2.0.18
